chore(NA): move maps plugin test fixtures out of __tests__ folder (#87764)
* chore(NA): move maps plugin test fixtures out of __tests__ folder

* chore(NA): last missing __tests__ folders renamed
