+++ x-pack/plugins/maps/public/classes/layers/__fixtures__/mock_sync_context.ts
