+++ x-pack/plugins/maps/public/classes/styles/vector/properties/dynamic_size_property.test.tsx
@@ -19,7 +19,7 @@ import { RawValue, VECTOR_STYLES } from '../../../../../common/constants';
 import { IField } from '../../../fields/field';
 import { Map as MbMap } from 'mapbox-gl';
 import { SizeDynamicOptions } from '../../../../../common/descriptor_types';
-import { mockField, MockLayer, MockStyle } from './__tests__/test_util';
+import { mockField, MockLayer, MockStyle } from './test_helpers/test_util';
 import { IVectorLayer } from '../../../layers/vector_layer/vector_layer';
 
 export class MockMbMap {
