+++ x-pack/plugins/maps/public/classes/styles/vector/properties/test_helpers/test_util.ts
