+++ x-pack/plugins/maps/public/classes/styles/vector/vector_style.test.js
@@ -12,7 +12,7 @@ import {
   VECTOR_SHAPE_TYPE,
   VECTOR_STYLES,
 } from '../../../../common/constants';
-import { MockField } from './properties/__tests__/test_util';
+import { MockField } from './properties/test_helpers/test_util';
 
 jest.mock('../../../kibana_services');
