+++ x-pack/plugins/maps/server/mvt/__fixtures__/json/0_0_0_gridagg.json
