+++ x-pack/plugins/maps/server/mvt/__fixtures__/tile_es_responses.ts
